+++ RELEASES.md
@@ -2,7 +2,7 @@
 
 > Note: ES3/IE8 compatibility will be removed in the future v3.x.x releases (scheduled for mid-late 2022), so if you need to retain ES3 compatibility you will need to remain on the 2.x.x versions of the SDK or your runtime will need install polyfill's to your ES3 environment before loading / initializing the SDK.
 
-## 3.4.0-beta (Unreleased)
+## 3.4.0-beta (February 23rd, 2026)
 
 ### Significant Changes